examples/measure_overhead: fix some warnings & ubsan errors

+++ examples/measure_overhead/measure_distributed_matmul.cc
@@ -77,7 +77,7 @@ void print_progress(int p, int total, int width = 80) {
 
 struct {
   int platform_index = -1;
-  int sample_count = 1;
+  size_t sample_count = 1;
   // This is the square root of the number of integers in the buffer, not bytes.
   size_t matrix_size = 128;
   int show_result = 0;
@@ -349,7 +349,7 @@ bool measure_platform(cl::Platform &platform, int index) {
         options.sample_count);
     std::vector<double> times(options.sample_count);
     print_progress(0, options.sample_count);
-    for (int i = 0; i < options.sample_count; ++i) {
+    for (size_t i = 0; i < options.sample_count; ++i) {
       using namespace std::chrono;
       starts[i] = steady_clock::now();
       run_iteration(kern, command_queues, buffers, partitions, events_by_device,
@@ -395,7 +395,7 @@ bool measure_platform(cl::Platform &platform, int index) {
       for (cl::CommandQueue &cq : command_queues)
         cq.finish();
       std::vector<double> device_times(options.sample_count);
-      for (int j = 0; j < events_by_device[i].size(); ++j) {
+      for (size_t j = 0; j < events_by_device[i].size(); ++j) {
         cl::Event &e = events_by_device[i][j];
         if (e.get() == nullptr)
           continue;
